requirements.txt
torch>=2.0.0
transformers>=4.25.0
sentence-transformers>=2.2.0
opencv-python>=4.8.0
faiss-cpu>=1.7.4
fastapi>=0.95.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
requests>=2.28.0
tqdm>=4.64.0
pillow>=9.0.0
pandas>=2.0.0
streamlit>=1.25.0